refactor: colocar 20 por defecto en el seed de test y mover la logica de crear curso con tipo de nota por defecto al usecase
