+++ app/src/main/java/com/app/grader/domain/usecase/course/SaveCourseUseCase.kt
@@ -1,5 +1,7 @@
 package com.app.grader.domain.usecase.course
 
+import com.app.grader.core.appConfig.toTypeGradeId
+import com.app.grader.data.appConfig.AppConfigRepository
 import com.app.grader.domain.model.CourseModel
 import com.app.grader.domain.model.Resource
 import com.app.grader.domain.repository.LocalStorageRepository
@@ -8,12 +10,17 @@ import kotlinx.coroutines.flow.channelFlow
 import javax.inject.Inject
 
 class SaveCourseUseCase @Inject constructor(
-    private val repository: LocalStorageRepository
+    private val repository: LocalStorageRepository,
+    private val appConfigRepository: AppConfigRepository
 ) {
     operator fun invoke(courseModel: CourseModel): Flow<Resource<Long>> = channelFlow {
         try {
             send(Resource.Loading())
-            val data = repository.saveCourse(courseModel)
+            val data = repository.saveCourse(
+                courseModel.copy(
+                    typeGradeId = resolveTypeGradeId(courseModel.typeGradeId)
+                )
+            )
             if (data.toInt() != -1){
                 send(
                     Resource.Success(data = data)
@@ -28,5 +35,10 @@ class SaveCourseUseCase @Inject constructor(
                 Resource.Error(e.message ?: "Unknown Error")
             )
         }
+    }
+
+    private fun resolveTypeGradeId(typeGradeId: Int): Int {
+        if (typeGradeId > 0) return typeGradeId
+        return appConfigRepository.getTypeGrade().toTypeGradeId()
     }
 }
